requirements.txt
django-phonenumber-field==8.0.0
phonenumberslite==8.13.43
environs==11.0.0
Pillow==10.4.0
djangorestframework==3.15.2